PHOENIX-628 Support native JSON data type - fix for first review
