+++ phoenix-core/src/main/java/org/apache/phoenix/expression/ExpressionType.java
@@ -43,8 +43,6 @@ import org.apache.phoenix.expression.function.FloorFunction;
 import org.apache.phoenix.expression.function.HourFunction;
 import org.apache.phoenix.expression.function.IndexStateNameFunction;
 import org.apache.phoenix.expression.function.InvertFunction;
-import org.apache.phoenix.expression.function.JsonExtractPathFunction;
-import org.apache.phoenix.expression.function.JsonExtractPathTextFunction;
 import org.apache.phoenix.expression.function.LTrimFunction;
 import org.apache.phoenix.expression.function.LastValueFunction;
 import org.apache.phoenix.expression.function.LengthFunction;
@@ -207,8 +205,6 @@ public enum ExpressionType {
     WeekFunction(WeekFunction.class),
     HourFunction(HourFunction.class),
     NowFunction(NowFunction.class),
-    JsonExtractPathFunction(JsonExtractPathFunction.class),
-    JsonExtractPathTextFunction(JsonExtractPathTextFunction.class)
     ;
 
     ExpressionType(Class<? extends Expression> clazz) {
