+++ phoenix-core/src/main/java/org/apache/phoenix/schema/types/PDataTypeFactory.java
@@ -96,7 +96,7 @@ public class PDataTypeFactory {
     types.add(PVarbinaryArray.INSTANCE);
     types.add(PVarchar.INSTANCE);
     types.add(PVarcharArray.INSTANCE);
-    types.add(PJsonDataType.INSTANCE);
+    types.add(PJson.INSTANCE);
 
     classToInstance = new HashMap<>(types.size());
     for (PDataType t : types) {
